Refactor rnpkeys to be usable in cpp CLI tests; Move code which is used only in rnpkeys to rnpkeys.cpp
