+++ src/rnpkeys/rnpkeys.h
@@ -4,7 +4,6 @@
 #include <stdbool.h>
 #include <sys/param.h>
 
-#define EXIT_ERROR 2
 #define DEFAULT_RSA_NUMBITS 2048
 
 typedef enum {
@@ -46,10 +45,10 @@ typedef enum {
 
 rnp_result_t rnp_generate_key_expert_mode(rnp_t *rnp, rnp_cfg_t *cfg);
 bool         rnp_cmd(rnp_cfg_t *cfg, rnp_t *rnp, optdefs_t cmd, char *f);
-int          setoption(rnp_cfg_t *cfg, optdefs_t *cmd, int val, char *arg);
+bool         setoption(rnp_cfg_t *cfg, optdefs_t *cmd, int val, char *arg);
 void         print_praise(void);
 void         print_usage(const char *usagemsg);
-int          parse_option(rnp_cfg_t *cfg, optdefs_t *cmd, const char *s);
+bool         parse_option(rnp_cfg_t *cfg, optdefs_t *cmd, const char *s);
 
 /* -----------------------------------------------------------------------------
  * @brief   Initializes rnpkeys. Function allocates memory dynamically for
